users can now delete a sub task with its associated list of items

+++ api/src/sub-task/sub-task.service.ts
@@ -3,16 +3,18 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { SubTask } from './sub-task.entity';
 import { CreateSubTaskDto } from './dto/createSubTaskDto';
+import { ItemService } from 'src/item/item.service';
 
 @Injectable()
 export class SubTaskService {
   constructor(
+    private itemService: ItemService,
     @InjectRepository(SubTask)
-    private subTaskRepository: Repository<SubTask>
+    private subTaskRepository: Repository<SubTask>,
   ) {}
 
   async getSubTasks(id: number) {
-    return await this.subTaskRepository.find({ where : { task: { id } }, relations: ['items', 'subTask.items'] })
+    return await this.subTaskRepository.find({ where : { task: { id } }, relations: ['items'] })
   }
 
   async createSubTask(createSubTaskDto: CreateSubTaskDto, taskId: number) {
@@ -31,9 +33,15 @@ export class SubTaskService {
     if (!subTaskToDelete) {
       throw new Error('sub task does not exist!')
     } else {
-      await this.subTaskRepository.remove(subTaskToDelete);
+      // const itemsBelongingToSubTask = await this.itemService.getSubTaskItems(subTaskId)
+      const itemsRemoved = await this.itemService.deleteAllItems(subTaskId);
+      if (itemsRemoved) {
+      const deleteRespsonse = await this.subTaskRepository.remove(subTaskToDelete);
       const subTasks = await this.getSubTasks(taskId)
       return subTasks;
+      } else {
+        throw new Error('Unable to remove all associated items, cannot delete subTask!')
+      }
     }
   }
